+++ src/main/java/roomescape/controller/dto/ReservationRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.controller.dto;
 
 public record ReservationRequest(
         String name,
