+++ src/main/java/roomescape/controller/dto/ReservationTimeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.controller.dto;
 
 public record ReservationTimeRequest(String startAt) {
 }
